// dashboard/src/utils/hasura-api/generated/schemas/createFunctionPermissionArgs.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { Function } from './function';

export interface CreateFunctionPermissionArgs {
  function: Function;
  /** Name of the role */
  role: string;
  /** Name of the source database of the function */
  source?: string;
}