// dashboard/src/utils/hasura-api/generated/schemas/createFunctionPermissionBulkOperation.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { CreateFunctionPermissionBulkOperationType } from './createFunctionPermissionBulkOperationType';
import type { CreateFunctionPermissionStep } from './createFunctionPermissionStep';

export interface CreateFunctionPermissionBulkOperation {
  type: CreateFunctionPermissionBulkOperationType;
  /** Source of the operation */
  source: string;
  /** Resource version of the operation */
  resource_version: number;
  args: CreateFunctionPermissionStep[];
}