// dashboard/src/utils/hasura-api/generated/schemas/dropFunctionPermissionBulkOperation.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { DropFunctionPermissionBulkOperationType } from './dropFunctionPermissionBulkOperationType';
import type { DropFunctionPermissionStep } from './dropFunctionPermissionStep';

export interface DropFunctionPermissionBulkOperation {
  type: DropFunctionPermissionBulkOperationType;
  /** Source of the operation */
  source: string;
  /** Resource version of the operation */
  resource_version: number;
  args: DropFunctionPermissionStep[];
}